requirements.txt
datasets==2.19.0
evaluate==0.1.0
faiss==1.7.1
h5py==3.11.0
huggingface-hub==0.22.2
nltk==3.6.5
numpy==1.24.4
pyserini==0.22.1
pytorch-lightning==2.2.2
pytrec-eval==0.5
sentence-transformers==2.2.2
sentencepiece==0.1.96
tokenizers==0.10.3
toml==0.10.2
torch==1.13.1
torchaudio==0.13.1
torchmetrics==1.3.2
torchvision==0.14.1
tqdm==4.66.2
transformers==4.12.5
trec-car-tools==2.6
wandb==0.16.6